requirements.txt
streamlit
langchain
langchain-openai
langchain-community
chromadb
pypdf
python-dotenv
loguru
tqdm
tiktoken